+++ src/pgs/image.rs
@@ -5,6 +5,8 @@ use bindings::Windows::Win32::System::WinRT::IBufferByteAccess;
 use bindings::Windows::UI::Color;
 use windows::Interface;
 
+use crate::interop::as_mut_slice;
+
 use super::types::ObjectDef;
 
 #[derive(Debug)]
@@ -66,11 +68,3 @@ pub fn decode_image(
     Ok(bitmap)
 }
 
-unsafe fn as_mut_slice(buffer: &Buffer) -> windows::Result<&mut [u8]> {
-    let interop = buffer.cast::<IBufferByteAccess>()?;
-    let mut data = std::ptr::null_mut();
-    let len = buffer.Length()?;
-
-    interop.Buffer(&mut data).ok()?;
-    Ok(std::slice::from_raw_parts_mut(data, len as _))
-}
